requirements.txt
# Data processing
pandas==1.5.3
numpy==1.24.3
pyarrow==12.0.0

# Cloud storage
google-cloud-storage==2.9.0
google-cloud-bigquery==3.11.0

# Data transformation
dbt-core==1.5.0
dbt-bigquery==1.5.0

# Data validation
great-expectations==0.16.13
setuptools==68.0.0

# Orchestration
apache-airflow==2.6.3
apache-airflow-providers-google==10.0.0
apache-airflow-providers-dbt-cloud==3.2.0
sqlalchemy==1.4.46
pendulum==2.1.2

# Testing
pytest==7.3.1
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0 